requirements.txt
django==4.1.5
argon2-cffi==21.3.0
django-environ==0.9.0
djangorestframework==3.14.0
pillow==9.4.0
psycopg2==2.9.5
psycopg2-binary==2.9.5

django-allauth==0.52.0
django-rest-auth==0.9.5
django-crispy-forms==1.14.0
crispy-tailwind==0.5.0
django-tailwind==3.4.0
markdown==3.4.1
django-countries
python-dateutil==2.8.2
channels==4.0.0
sqlparse==0.4.3
asgiref==3.6.0
pytz==2022.7.1
mutagen==1.46.0
django-naomi

pandas~=1.5.3
numpy~=1.24.2
openpyxl~=3.1.0
phonenumbers~=8.13.7
requests
django-paypal
